+++ R/LF_to_HA.R
@@ -6,7 +6,7 @@
 #' @export
 #'
 LF_to_HA <- function(data_agro_LF){
-  if(!is(data_agro_LF, "data_agro_LF")){ stop(substitute(data_agro_LF), " must be formated with type = \"data_agro_LA\", see PPBstats::format_data_PPBstats().") }
+  if(!is(data_agro_LF, "data_agro_LF")){ stop(substitute(data_agro_LF), " must be formated with type = \"data_agro_LF\", see PPBstats::format_data_PPBstats().") }
   data_agro_HA = data_agro_LF
   data_agro_HA$version = as.character(data_agro_HA$version)
   data_agro_HA$version[which(data_agro_HA$version == "local")] = "home"