+++ R/PPBstats-package.R
@@ -64,9 +64,9 @@ NULL
 #' Data to study home and away
 #'
 #' @description
-#' A dataset containing values for the following factors: seed_lot, location, year, germplasm, block, X, Y, group and version, Y1, Y1$date, y2, y2$date, y3, y3$date.
-#' group represents home or away
-#' version represents the location where come the germplasm from.
+#' A dataset containing values for the following factors: seed_lot, location, year, germplasm, block, X, Y, origin and version, Y1, Y1$date, y2, y2$date, y3, y3$date.
+#' origin represents the location where come the germplasm from.
+#' version  represents home or away.
 #'
 #' @usage
 #' data(data_agro_HA)
@@ -82,9 +82,9 @@ NULL
 #' Data to study local and foreign
 #'
 #' @description
-#' A dataset containing values for the following factors: seed_lot, location, year, germplasm, block, X, Y, group and version, Y1, Y1$date, y2, y2$date, y3, y3$date.
-#' group represents local or foreign
-#' version represents the location where come the germplasm from.
+#' A dataset containing values for the following factors: seed_lot, location, year, germplasm, block, X, Y, origin and version, Y1, Y1$date, y2, y2$date, y3, y3$date.
+#' origin represents the location where come the germplasm from.
+#' version represents local or foreign.
 #'
 #' @usage
 #' data(data_agro_LF)