+++ NEWS.md
@@ -33,6 +33,7 @@
     - model_local_foreign
 - update ggradar in common_function
 - design_experiment, cf #39
+- home away and local foreign workflow: change column name group to origin to be more clear
 
 ### add new functions
 - post_hoc_variation, cf #141